requirements.txt
asgiref==3.6.0
backports.zoneinfo==0.2.1
Django==4.1.5
djangorestframework==3.14.0
importlib-metadata==6.0.0
Markdown==3.4.1
pytz==2022.7.1
sqlparse==0.4.3
zipp==3.11.0
Faker==16.6.0
django-filter==22.1
django-heroku==0.3.1
gunicorn==20.0.4



